Add venue controller and filter by user id, also index on that column

+++ database/migrations/48_roadmap_point_venue_form_detail.js
@@ -23,7 +23,7 @@ class RoadmapPointVenueFormDetailSchema extends Schema {
         .integer('item_id')
         .references('id')
         .inTable(itemTableName)
-      table.float('qty')
+      table.decimal('qty')
       table.text('notes')
       table.jsonb('metadata')
       table.timestamps()
